独立出 hooks 部分并添加注释
